# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Comfort taxi API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/driver": {
            "get": {
                "tags": [
                    "driver"
                ],
                "summary": "Get list of drivers",
                "operationId": "91ce118ace6bf8c1618dae381e22e051",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "400": {
                        "description": "drivers isn't found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/driver/{id}": {
            "get": {
                "tags": [
                    "driver"
                ],
                "summary": "Get driver by id",
                "description": "Get driver by id",
                "operationId": "getdriver",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of driver",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "driver": {
                                            "$ref": "#/components/schemas/DriverResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "driver is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "driver is not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "driver"
                ],
                "summary": "Create driver",
                "description": "Create new driver",
                "operationId": "createdriver",
                "requestBody": {
                    "description": "Send data to create a new driver",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user": {
                                        "$ref": "#/components/schemas/DriverCreateRequest"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "task": {
                                            "$ref": "#/components/schemas/DriverResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error when creating driver",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error when creating driver"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "driver"
                ],
                "summary": "Delete driver by id",
                "description": "Delete driver by id",
                "operationId": "deletedriver",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of driver",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "driver": {
                                            "$ref": "#/components/schemas/DriverResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "driver is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "driver is not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "driver"
                ],
                "summary": "Update driver",
                "description": "Update driver",
                "operationId": "updatedriver",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of driver",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Send data to update a driver",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "driver": {
                                        "$ref": "#/components/schemas/DriverUpdateRequest"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "driver": {
                                            "$ref": "#/components/schemas/DriverResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "driver is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "driver is not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error when updating driver",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error when updating driver"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DriverCreateRequest": {
                "required": [
                    "firstname",
                    "secondname",
                    "date_hired",
                    "driving_licence_id",
                    "rating"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "secondname": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "Иванов"
                    },
                    "firstname": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "Иван"
                    },
                    "patronymic": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "Иванович"
                    },
                    "date_hired": {
                        "type": "string",
                        "format": "date",
                        "readOnly": "true",
                        "example": "2019-02-25"
                    },
                    "date_fired": {
                        "type": "string",
                        "format": "date",
                        "readOnly": "true",
                        "example": "2019-02-25"
                    },
                    "rating": {
                        "type": "number",
                        "format": "double",
                        "readOnly": "true",
                        "example": "4.8"
                    },
                    "driving_licence_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "DriverCreateRequest"
                }
            },
...
            "DriverUpdateRequest": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "secondname": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "Иванов"
                    },
                    "firstname": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "Иван"
                    },
                    "patronymic": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "Иванович"
                    },
                    "date_hired": {
                        "type": "string",
                        "format": "date",
                        "readOnly": "true",
                        "example": "2019-02-25"
                    },
                    "date_fired": {
                        "type": "string",
                        "format": "date",
                        "readOnly": "true",
                        "example": "2019-02-25"
                    },
                    "rating": {
                        "type": "number",
                        "format": "double",
                        "readOnly": "true",
                        "example": "4.8"
                    },
                    "driving_licence_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "DriverUpdateRequest"
                }
            },
...
            "DriverResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "secondname": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "Иванов"
                    },
                    "firstname": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "Иван"
                    },
                    "patronymic": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "Иванович"
                    },
                    "date_hired": {
                        "type": "string",
                        "format": "date",
                        "readOnly": "true",
                        "example": "2019-02-25"
                    },
                    "date_fired": {
                        "type": "string",
                        "format": "date",
                        "readOnly": "true",
                        "example": "2019-02-25"
                    },
                    "rating": {
                        "type": "number",
                        "format": "double",
                        "readOnly": "true",
                        "example": "4.8"
                    },
                    "driving_licence_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "DriverResource"
                }
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    ]
}